documentation: reflections for sessions 3 and 4.
Thinking about ways to improve the course in the future, specifically making more of a consequence for not reading. Need to create a proper check to make sure this feels necessary. Simply saying that it will make you understand the material more clearly is not enough.

Also reflecting on this idea of "asynchronous" class sessions. An interesting one to unpack.